server/requirements.txt
algoliasearch-django>=2.0,<3.0
django>=4.2.0,<4.3.0
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers
Pillow
python-dotenv
newsapi-python
nltk>=3.8.1,<3.9.0